+++ lib/screens/quds_popup_menu.dart
@@ -263,7 +263,7 @@ class _QudsPopupMenuState extends State<_QudsPopupMenu> {
                                 sections.last.titleText,
                                 style: Theme.of(context)
                                     .textTheme
-                                    .headline5!
+                                    .headlineSmall!
                                     .copyWith(fontWeight: FontWeight.bold),
                               )
                             ],